bug fix: change type of courseNum in KDAMCHECK, encoderdecoder,student

+++ spl-net/src/main/java/bgu/spl/net/impl/RegistrationSystem/Student.java
@@ -18,7 +18,7 @@ public class Student extends User {
     public boolean haveAllKdamCourses(Database database,Course course){
         ArrayList<String> KdamCourses = course.getKdamCoursesList();
         for (String num:KdamCourses){
-            Course kdamCourse= database.getCourses().get(Integer.parseInt(num));
+            Course kdamCourse= database.getCourses().get(Short.parseShort(num));
             if (!registeredCourses.contains(kdamCourse)){
                 return false;
             }
